Fixed `VariationOfFileSize` trigger
There was 2 bugs:

1. If more than 1 task was using this trigger, each of them would have a
constant activation due that the variable used to store the
last size of the file (`previousFileSize`) would be shared between tasks.
That was fixed using a map with the ID of the task as the key for
access/modify the values.
2. Instead of activate the trigger when the size variated, it would be
activated when the size still with the same value (on line 67 to be more
precise). That was fixed too.