Removed setting `WebUI.RequireCredentials`
Basically there is no place where to use it. First: I think that PW must
have their users identified, preventing problems caused by someone
that you don't even know. Also, not using credentials is an open door to the
info in the device that is running it, something that obviously
must be totally avoided. Second: the WebUI is nothing if the APIs are
not enabled (plus to the local access setting, which basically closes
the access to anyone outside your computer).